console: Refactor collaborator form
Move from class components to functional components
